UnixTools: Fix Python bots launch ##63
The Python bots need the Python interpreter to run and we form
a commandline without absolute or relative path (.e.g. python
bots/bot/run.py). It doesn't work with the current implementation as
don't take PATH into account.

This commit uses execvp for the Python bots to fix the issue. For the
binary bots the old approach is preserved in order to avoid
path specifications for binary executables (e.g. ./Bot) otherwise we
have to add the executables to PATH which is not convenient.