Use sc2::TerminateProcess to kill SC2 clients
The API tracks all the spawned game processes when one calls sc2::StartProcess.
(seems that it is done to kill all the processes spawned by the game).
It is expected that the API's user will call sc2::TerminateProcess which will
kill all the childs and free remembered pids. Since we use our
own 'killSc2' implementation the pid is never freed and we have a kind
of memory leak.

Lets use sc2::TerminateProcess to avoid the leak.

Also some renaming done for better readability.